requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
scikit-learn
catboost
xgboost
setuptools
datetime
pymysql
python-dotenv
mysql-connector-python
scikit-learn
